Add references to existing snippets from spannerconnection members.  This ensures most of the non-ADO.NET methods have snippets. Moved the overviews to SpannerConnection snippets so they could more easily be referenced via 'Additional'
